# feat: persistent cache snapshot save file hash by default