--- a/packages/rspack-test-tools/tests/cacheCases/common/update-file/file.js
+++ b/packages/rspack-test-tools/tests/cacheCases/common/update-file/file.js
@@ -1,0 +1,9 @@
+export default 1;
+---
+export default 2;
+---
+export default 2;
+---
+<force_write>export default 2;
+---
+export default 3;