--- a/packages/rspack-test-tools/tests/cacheCases/common/update-file/loader.js
+++ b/packages/rspack-test-tools/tests/cacheCases/common/update-file/loader.js
@@ -1,0 +1,5 @@
+module.exports = function (content) {
+	const options = this.getOptions();
+	options.files.push(this.resourcePath);
+	return content;
+};