… packages/rspack-test-tools/tests/cacheCases/common/update-file/rspack.config.js