<!--
  Thank you for submitting a pull request!

  We appreciate the time and effort you have invested in making these changes. Please ensure that you provide enough information to allow others to review your pull request.

  Upon submission, your pull request will be automatically assigned with reviewers.

  If you want to learn more about contributing to this project, please visit: https://github.com/web-infra-dev/rspack/blob/main/CONTRIBUTING.md.
-->

## Summary

The persistent cache snapshot will store the compile times and use them to calculate which files were modified when hot start. However, this strategy is invalid in scenarios such as CI because the update time of all files will be refreshed.

This PR will make snapshot store both compile time and file hash, and when calculate modified files the file hash will be compared when the file modification time has been updated.

This PR will introduce some performance regression due to the addition of hash calculations, and I manually benchmarked the performance of snapshot generation in cold start.

| Name | File Count | Hash Count | Before | After |
| -- | -- | -- | -- | -- |
| Docusaurus | 4059 | 2143 | 22.32ms | 73.274ms |
| Large real project | 44905 | 8891 | 106.189ms | 365.41ms |

As can be seen from the table above, this takes three times longer than before, and the main time consumption occurs in file IO. This is currently acceptable in large real projects.

PS. For hot start, it will take 3 times longer if you update a lot of files like in CI, otherwise it will be the same as before. HMR will create snapshots incrementally so the performance impact is smaller.

<!-- Can you explain the reasoning behind implementing this change? What problem or issue does this pull request resolve? -->

<!-- It would be helpful if you could provide any relevant context, such as GitHub issues or related discussions. -->

## Checklist

<!--- Check and mark with an "x" -->

- [ ] Tests updated (or not required).
- [ ] Documentation updated (or not required).