@LingyuCoder 2025-04-17
can this issue https://github.com/web-infra-dev/rspack/issues/9064 be closed?