@jerrykingxyz 2025-04-17
> can this issue #9064 be closed?

No, this is another bug caused by storage invalidation, not snapshot invalidation. I will fix it in another PR.